[rpc][api] Remove TestCase, TestCaseRun and TestRun .get_bugs()
instead use the new TestCaseBug.filter methods with `case` or
`case_run` query attributes.

To get all bugs for a TestRun, first use TestRun.get_cases()
and then fetch all bugs with either the case_id or the case_run_id
query parameter.
