+++ docs/source/modules/tcms.xmlrpc.api.testcasebug.rst
@@ -1,0 +1,7 @@
+tcms\.xmlrpc\.api\.testcasebug module
+=====================================
+
+.. automodule:: tcms.xmlrpc.api.testcasebug
+    :members:
+    :undoc-members:
+    :show-inheritance:
