+++ tcms/settings/common.py
@@ -264,6 +264,7 @@ MODERNRPC_METHODS_MODULES = [
         'tcms.xmlrpc.api.priority',
         'tcms.xmlrpc.api.product',
         'tcms.xmlrpc.api.tag',
+        'tcms.xmlrpc.api.testcasebug',
         'tcms.xmlrpc.api.testcase',
         'tcms.xmlrpc.api.testcaserun',
         'tcms.xmlrpc.api.testcasestatus',
